- fixing inheritance initializers - using trade instance throughout paper exchange - adding event listener to trade manaher - moving list_trades and get_trade to trade manager - fixing trade manager pnl int to float - fixing woo exchange tests - adding trade manager event listener test

+++ src/superalgorithm/types/data_types.py
@@ -140,7 +140,7 @@ class Trade:
     price: float
     quantity: float
     server_order_id: str
-    pnl: int = 0
+    pnl: float = 0.0
 
     def to_dict(self):
         data_dict = asdict(self)
